[NFC] Move SimplifyCFG class to its own header so that it can be accessed by the unit test infra
